Battery-Database/tests/parsers/test_parser.py
import logging

from nomad.datamodel import EntryArchive

from battery_database.parsers.parser import NewParser


def test_parse_file():
    parser = NewParser()
    archive = EntryArchive()
    parser.parse('tests/data/example.out', archive, logging.getLogger())

    assert archive.workflow2.name == 'test'
